Restructure multi-ledger test
